+++ CHANGELOG.md
@@ -3,7 +3,10 @@
 * Moved all common code used by all controllers
   to a base template io controller class.
 * Added an SDL2IO controllable which replaces
-  the old SdlIoController. 
+  the old SdlIoController.
+* Added an empty io controllable called `BlankIO`
+  which can be used as a template for future
+  io controllable implementations.
 
 1.1.0 [26/12/25]
 * Added disabled GitHub Actions CI/CD support.